Revert "VPN integration"
This reverts commit 19f86eb0801363bef5cc485b28d861b835bdfc25.
