+++ pkg/cli/cmds/agent.go
@@ -32,8 +32,6 @@ type Agent struct {
 	FlannelIface             string
 	FlannelConf              string
 	FlannelCniConfFile       string
-	VPNAuth                  string
-	VPNAuthFile              string
 	Debug                    bool
 	Rootless                 bool
 	RootlessAlreadyUnshared  bool
@@ -154,18 +152,6 @@ var (
 		Name:        "flannel-cni-conf",
 		Usage:       "(agent/networking) Override default flannel cni config file",
 		Destination: &AgentConfig.FlannelCniConfFile,
-	}
-	VPNAuth = &cli.StringFlag{
-		Name:        "vpn-auth",
-		Usage:       "(agent/networking) (experimental) Credentials for the VPN provider. It must include the provider name and join key in the format name=<vpn-provider>,joinKey=<key>",
-		EnvVar:      version.ProgramUpper + "_VPN_AUTH",
-		Destination: &AgentConfig.VPNAuth,
-	}
-	VPNAuthFile = &cli.StringFlag{
-		Name:        "vpn-auth-file",
-		Usage:       "(agent/networking) (experimental) File containing credentials for the VPN provider. It must include the provider name and join key in the format name=<vpn-provider>,joinKey=<key>",
-		EnvVar:      version.ProgramUpper + "_VPN_AUTH_FILE",
-		Destination: &AgentConfig.VPNAuthFile,
 	}
 	ResolvConfFlag = &cli.StringFlag{
 		Name:        "resolv-conf",
@@ -301,8 +287,6 @@ func NewAgentCommand(action func(ctx *cli.Context) error) cli.Command {
 				EnvVar:      version.ProgramUpper + "_CLUSTER_SECRET",
 				Hidden:      true,
 			},
-			VPNAuth,
-			VPNAuthFile,
 		},
 	}
 }
