+++ pkg/passwd/passwd.go
@@ -7,6 +7,7 @@ import (
 	"os"
 	"strings"
 
+	"github.com/k3s-io/k3s/pkg/token"
 	"github.com/k3s-io/k3s/pkg/util"
 )
 
@@ -82,7 +83,7 @@ func (p *Passwd) EnsureUser(name, role, passwd string) error {
 	}
 
 	if passwd == "" {
-		token, err := util.Random(16)
+		token, err := token.Random(16)
 		if err != nil {
 			return err
 		}
